+++ libs/components/homepage/TrendPropertyCard.tsx
@@ -34,16 +34,16 @@ const TrendPropertyCard = (props: TrendPropertyCardProps) => {
 				<Box
 					component={'div'}
 					className={'card-img'}
-					style={{ backgroundImage: `url(${REACT_APP_API_URL}/${property?.propertyImages[0]})` }}
+					style={{ backgroundImage: `url(${REACT_APP_API_URL}/${property?.productImages[0]})` }}
 					onClick={() => {pushDetailHandler(property._id)}}
 				>
-					<div>${property.propertyPrice}</div>
+					<div>${property.productPrice}</div>
 				</Box>
 				<Box component={'div'} className={'info'}>
 					<strong className={'title'}
 					onClick={() => {pushDetailHandler(property._id)}}
-					>{property.propertyTitle}</strong>
-					<p className={'desc'}>{property.propertyDesc ?? 'no description'}</p>
+					>{property.productName}</strong>
+					<p className={'desc'}>{property.productDesc ?? 'no description'}</p>
 					<div className={'options'}>
 						<div>
 							<img src="/img/icons/bed.svg" alt="" />
@@ -68,7 +68,7 @@ const TrendPropertyCard = (props: TrendPropertyCardProps) => {
 							<IconButton color={'default'}>
 								<RemoveRedEyeIcon />
 							</IconButton>
-							<Typography className="view-cnt">{property?.propertyViews}</Typography>
+							<Typography className="view-cnt">{property?.productViews}</Typography>
 							<IconButton color={'default'} onClick={() => likePropertyHandler(user, property?._id)}>
 								{property?.meLiked && property?.meLiked[0]?.myFavorite ? (
 									<FavoriteIcon style={{ color: 'red' }} />
@@ -76,7 +76,7 @@ const TrendPropertyCard = (props: TrendPropertyCardProps) => {
 									<FavoriteIcon />
 								)}
 							</IconButton>
-							<Typography className="view-cnt">{property?.propertyLikes}</Typography>
+							<Typography className="view-cnt">{property?.productLikes}</Typography>
 						</div>
 					</div>
 				</Box>
@@ -88,23 +88,23 @@ const TrendPropertyCard = (props: TrendPropertyCardProps) => {
 				<Box
 					component={'div'}
 					className={'card-img'}
-					//style={{ backgroundImage: `url(${REACT_APP_API_URL}/${property?.propertyImages[0]})` }}
+					//style={{ backgroundImage: `url(${REACT_APP_API_URL}/${property?.productImages[0]})` }}
 					onClick={() => {pushDetailHandler(property._id)}}
 				>  
 				 	
-                <img src={`${REACT_APP_API_URL}/${property?.propertyImages[0]}`} alt="Product Image" />
+                <img src={`${REACT_APP_API_URL}/${property?.productImages[0]}`} alt="Product Image" />
            
 				</Box>
 			<Box component={'div'} className={'info'}>
 				{/* 	<strong className={'title'}
 					onClick={() => {pushDetailHandler(property._id)}}
-					>{property.propertyTitle}</strong>
-					<p className={'desc'}>{property.propertyDesc ?? 'no description'}</p>
+					>{property.productName}</strong>
+					<p className={'desc'}>{property.productDesc ?? 'no description'}</p>
 					
 					<Divider sx={{ mt: '15px', mb: '17px' }} /> */}
 					<strong className={'title'}
 					onClick={() => {pushDetailHandler(property._id)}}
-					>{property.propertyTitle}</strong>
+					>{property.productName}</strong>
 					
 				</Box> 
 			</Stack>